SG-964 SG-950 Fix shooting with mutliple guns
Hardcoding values is not acceptable, this just wasted me 3 hours of my time because you 'thought' it worked. No it didn't work, you looped through all the players and returned an array, why the hell would you think it is right to use the first element of that array instead of using playerId.